hybrid_rag_agent/requirements.txt
pydantic-ai>=0.0.10
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
openai>=1.0.0
numpy>=1.24.0
rich>=13.0.0
asyncio>=3.4.3